moving schema package to commons
