+++ hermes-common/build.gradle
@@ -14,6 +14,9 @@ dependencies {
         exclude module: 'slf4j-log4j12'
         exclude module: 'log4j'
     }
+    compile (group: 'com.github.fge', name: 'json-schema-validator', version: '2.2.6') {
+        exclude group: 'net.sf.jopt-simple'
+    }
 
     compile group: 'org.glassfish.hk2', name: 'hk2-locator', version: '2.3.0-b01'
     compile group: 'org.glassfish.jersey.core', name: 'jersey-client', version: versions.jersey
