+++ hermes-common/src/main/java/pl/allegro/tech/hermes/common/schema/MessageSchemaCompiler.java
@@ -1,4 +1,4 @@
-package pl.allegro.tech.hermes.frontend.schema;
+package pl.allegro.tech.hermes.common.schema;
 
 @FunctionalInterface
 public interface MessageSchemaCompiler<T> {
