+++ hermes-common/src/main/java/pl/allegro/tech/hermes/common/schema/TopicFieldMessageSchemaSourceRepository.java
@@ -1,4 +1,4 @@
-package pl.allegro.tech.hermes.frontend.schema;
+package pl.allegro.tech.hermes.common.schema;
 
 import pl.allegro.tech.hermes.api.Topic;
 
@@ -8,5 +8,5 @@ public class TopicFieldMessageSchemaSourceRepository implements MessageSchemaSou
     public String getSchemaSource(Topic topic) {
         return topic.getMessageSchema();
     }
-    
+
 }
