+++ hermes-frontend/build.gradle
@@ -3,9 +3,6 @@ dependencies {
     compile project(':hermes-tracker')
     compile project(':hermes-metrics')
 
-    compile (group: 'com.github.fge', name: 'json-schema-validator', version: '2.2.6') {
-        exclude group: 'net.sf.jopt-simple'
-    }
     compile group: 'io.undertow', name: 'undertow-servlet', version: '1.2.7.Final'
     compile group: 'org.apache.kafka', name: 'kafka-clients', version: '0.8.2.0'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
