+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/di/FrontendBinder.java
@@ -5,6 +5,11 @@ import org.I0Itec.zkclient.ZkClient;
 import org.apache.avro.Schema;
 import org.glassfish.hk2.api.TypeLiteral;
 import org.glassfish.hk2.utilities.binding.AbstractBinder;
+import pl.allegro.tech.hermes.common.schema.AvroMessageSchemaRepositoryFactory;
+import pl.allegro.tech.hermes.common.schema.JsonMessageSchemaRepositoryFactory;
+import pl.allegro.tech.hermes.common.schema.MessageSchemaRepository;
+import pl.allegro.tech.hermes.common.schema.MessageSchemaSourceRepository;
+import pl.allegro.tech.hermes.common.schema.TopicFieldMessageSchemaSourceRepository;
 import pl.allegro.tech.hermes.frontend.cache.topic.TopicsCache;
 import pl.allegro.tech.hermes.frontend.cache.topic.zookeeper.ZookeeperTopicsCacheFactory;
 import pl.allegro.tech.hermes.frontend.producer.BrokerMessageProducer;
@@ -14,7 +19,6 @@ import pl.allegro.tech.hermes.frontend.producer.kafka.Producers;
 import pl.allegro.tech.hermes.frontend.publishing.MessageContentTypeEnforcer;
 import pl.allegro.tech.hermes.frontend.publishing.MessagePublisher;
 import pl.allegro.tech.hermes.frontend.publishing.PublishingServlet;
-import pl.allegro.tech.hermes.frontend.schema.*;
 import pl.allegro.tech.hermes.frontend.server.HermesServer;
 import pl.allegro.tech.hermes.frontend.services.HealthCheckService;
 import pl.allegro.tech.hermes.frontend.validator.*;
